fix these compiler warnings:
```
ls_calibration.ino: In function 'int32_t calculateReferenceX(byte)':
linnstrument-firmware.ino:1114:40: warning: left shift of negative value [-Wshift-negative-value]
 1114 | #define FXD_FROM_INT(a)  (int32_t)((a) << FXD_FBITS)
      |                                        ^
ls_calibration.ino:98:20: note: in expansion of macro 'FXD_FROM_INT'
   98 |     return FXD_MUL(FXD_FROM_INT(-1), FXD_CALX_HALF_UNIT) + FXD_CALX_BORDER_OFFSET;
      |                    ^~~~~~~~~~~~
ls_calibration.ino: In function 'boolean validateAndHealCalibrationData()':
linnstrument-firmware.ino:1114:40: warning: left shift of negative value [-Wshift-negative-value]
 1114 | #define FXD_FROM_INT(a)  (int32_t)((a) << FXD_FBITS)
      |                                        ^
ls_calibration.ino:289:35: note: in expansion of macro 'FXD_FROM_INT'
  289 |     int32_t previous_measured_x = FXD_FROM_INT(-500);
      |                                   ^~~~~~~~~~~~
```

+++ linnstrument-firmware.ino
@@ -1111,7 +1111,7 @@ constexpr const int SINGLE_PROJECT_SIZE = alignToWord32Boundary(sizeof(Sequencer
 /**************************************** FIXED POINT MATH ***************************************/
 
 #define FXD_FBITS        8
-#define FXD_FROM_INT(a)  (int32_t)((a) << FXD_FBITS)
+#define FXD_FROM_INT(a)  (int32_t)(((uint32_t)(a)) << FXD_FBITS)
 #define FXD_MAKE(a)      (int32_t)((a * (1 << FXD_FBITS)))
 
 inline int FXD_TO_INT(int32_t a) {
